refactor(react-grab): consolidate scan modules into scanner.ts; trim comments
Merges scan-controller and scan-trace-recorder into core/scanner.ts
(recorder -> scanner -> controller), shrinking the public surface to
createScanController/ScanController. Also trims the scan feature's
comments down to only non-obvious "why" notes per AGENTS.
