+++ packages/react-grab/src/core/scanner.ts
@@ -3,13 +3,21 @@ import {
   didFiberRender,
   getDisplayName,
   getNearestHostFiber,
+  getTimings,
   isCompositeFiber,
   isInstrumentationActive,
   traverseRenderedFibers,
   type Fiber,
   type FiberRoot,
 } from "bippy";
+import { createSignal, onCleanup, type Accessor } from "solid-js";
 import {
+  FEEDBACK_DURATION_MS,
+  LONG_ANIMATION_FRAME_ENTRY_TYPE,
+  MAX_SCAN_TRACE_COMMITS,
+  MAX_SCAN_TRACE_FIBERS_PER_COMMIT,
+  MAX_SCAN_TRACE_LOAF_ENTRIES,
+  MAX_SCAN_TRACE_LOAF_SCRIPTS,
   MIN_DEVICE_PIXEL_RATIO,
   OVERLAY_BORDER_COLOR_DEFAULT,
   OVERLAY_FILL_COLOR_DEFAULT,
@@ -20,25 +28,181 @@ import {
   SCAN_OUTLINE_DURATION_MS,
   Z_INDEX_SCAN_CANVAS,
 } from "../constants.js";
-import type { ScanTrace } from "../types.js";
-import { createScanTraceRecorder } from "./scan-trace-recorder.js";
-import { hideFromThirdParties } from "../utils/hide-from-third-parties.js";
-import { nativeCancelAnimationFrame, nativeRequestAnimationFrame } from "../utils/native-raf.js";
-import { isUsefulComponentName } from "../utils/is-useful-component-name.js";
+import type {
+  ScanCommit,
+  ScanLoafScript,
+  ScanLongAnimationFrame,
+  ScanRenderedFiber,
+  ScanTrace,
+} from "../types.js";
+import { copyContent } from "../utils/copy-content.js";
 import { createElementBounds } from "../utils/create-element-bounds.js";
 import { drawRoundedRectangle } from "../utils/draw-rounded-rectangle.js";
-import { parseBorderRadiusValue } from "../utils/parse-border-radius-value.js";
 import { formatRenderLabel } from "../utils/format-render-label.js";
+import { getChangeDescription } from "../utils/get-change-description.js";
+import { getFiberSource } from "../utils/get-fiber-source.js";
+import { hideFromThirdParties } from "../utils/hide-from-third-parties.js";
+import { isUsefulComponentName } from "../utils/is-useful-component-name.js";
+import { nativeCancelAnimationFrame, nativeRequestAnimationFrame } from "../utils/native-raf.js";
+import { parseBorderRadiusValue } from "../utils/parse-border-radius-value.js";
+import { serializeScanTrace } from "../utils/serialize-scan-trace.js";
+
+// ---------------------------------------------------------------------------
+// Trace recorder: per-commit fiber detail (actualDuration, why, where) + LoAFs.
+// ---------------------------------------------------------------------------
+
+// `long-animation-frame` fields aren't in the DOM lib types yet.
+interface LongAnimationFrameScriptTiming {
+  sourceURL?: string;
+  sourceFunctionName?: string;
+  sourceCharPosition?: number;
+  duration?: number;
+  forcedStyleAndLayoutDuration?: number;
+}
+
+interface LongAnimationFrameTiming extends PerformanceEntry {
+  blockingDuration?: number;
+  firstUIEventTimestamp?: number;
+  scripts?: LongAnimationFrameScriptTiming[];
+}
+
+interface CollectedFiber {
+  fiber: Fiber;
+  name: string;
+  actualDurationMs: number;
+  parentRendered: boolean;
+}
+
+interface ScanTraceRecorder {
+  begin: () => void;
+  end: () => void;
+  beginCommit: (timestamp: number) => void;
+  collectFiber: (fiber: Fiber, name: string, parentRendered: boolean) => void;
+  endCommit: () => void;
+  takeTrace: () => ScanTrace | null;
+}
+
+const isLongAnimationFrameSupported = (): boolean =>
+  typeof PerformanceObserver !== "undefined" &&
+  Boolean(PerformanceObserver.supportedEntryTypes?.includes(LONG_ANIMATION_FRAME_ENTRY_TYPE));
+
+const createScanTraceRecorder = (): ScanTraceRecorder => {
+  const commits: ScanCommit[] = [];
+  const longAnimationFrames: ScanLongAnimationFrame[] = [];
+  const collected: CollectedFiber[] = [];
+  let commitCount = 0;
+  let currentCommitTimestamp = 0;
+  let scanStartTimestamp = 0;
+  let observer: PerformanceObserver | null = null;
+
+  const ingestLoafEntries = (entries: PerformanceEntryList): void => {
+    for (const entry of entries) {
+      if (longAnimationFrames.length >= MAX_SCAN_TRACE_LOAF_ENTRIES) break;
+      const loafEntry = entry as LongAnimationFrameTiming;
+      const scripts: ScanLoafScript[] = (loafEntry.scripts ?? [])
+        .slice(0, MAX_SCAN_TRACE_LOAF_SCRIPTS)
+        .map((script) => ({
+          sourceURL: script.sourceURL ?? "",
+          sourceFunctionName: script.sourceFunctionName ?? "",
+          sourceCharPosition: script.sourceCharPosition ?? 0,
+          durationMs: script.duration ?? 0,
+          forcedStyleAndLayoutDurationMs: script.forcedStyleAndLayoutDuration ?? 0,
+        }));
+      longAnimationFrames.push({
+        startTimeMs: loafEntry.startTime,
+        durationMs: loafEntry.duration,
+        blockingDurationMs: loafEntry.blockingDuration ?? 0,
+        firstUIEventTimestampMs: loafEntry.firstUIEventTimestamp ?? 0,
+        scripts,
+      });
+    }
+  };
+
+  const begin = (): void => {
+    commits.length = 0;
+    longAnimationFrames.length = 0;
+    collected.length = 0;
+    commitCount = 0;
+    scanStartTimestamp = performance.now();
+    if (!isLongAnimationFrameSupported()) return;
+    observer = new PerformanceObserver((list) => ingestLoafEntries(list.getEntries()));
+    try {
+      observer.observe({ type: LONG_ANIMATION_FRAME_ENTRY_TYPE, buffered: false });
+    } catch {
+      observer = null;
+    }
+  };
+
+  const end = (): void => {
+    if (!observer) return;
+    // Drain buffered-but-undelivered entries (delivery is async) before disconnect.
+    ingestLoafEntries(observer.takeRecords());
+    observer.disconnect();
+    observer = null;
+  };
+
+  const beginCommit = (timestamp: number): void => {
+    currentCommitTimestamp = timestamp;
+    commitCount += 1;
+    collected.length = 0;
+  };
+
+  const collectFiber = (fiber: Fiber, name: string, parentRendered: boolean): void => {
+    collected.push({ fiber, name, actualDurationMs: fiber.actualDuration ?? 0, parentRendered });
+  };
+
+  const endCommit = (): void => {
+    if (collected.length === 0) return;
+    collected.sort((first, second) => second.actualDurationMs - first.actualDurationMs);
+
+    // actualDuration nests children, so the slowest fiber's is the commit total.
+    const totalActualDurationMs = collected[0].actualDurationMs;
+
+    // Enrich only the kept fibers, and now, while they're still this commit's
+    // live alternates.
+    const fibers: ScanRenderedFiber[] = collected
+      .slice(0, MAX_SCAN_TRACE_FIBERS_PER_COMMIT)
+      .map((entry) => ({
+        name: entry.name,
+        selfDurationMs: getTimings(entry.fiber).selfTime,
+        source: getFiberSource(entry.fiber),
+        change: getChangeDescription(entry.fiber, entry.parentRendered),
+      }));
+
+    commits.push({
+      timestampMs: currentCommitTimestamp,
+      totalActualDurationMs,
+      renderedFiberCount: collected.length,
+      fibers,
+    });
+    if (commits.length > MAX_SCAN_TRACE_COMMITS) commits.shift();
+    collected.length = 0;
+  };
+
+  const takeTrace = (): ScanTrace | null => {
+    if (commits.length === 0 && longAnimationFrames.length === 0) return null;
+    return {
+      durationMs: performance.now() - scanStartTimestamp,
+      commitCount,
+      commits: commits.slice(),
+      longAnimationFrames: longAnimationFrames.slice(),
+    };
+  };
+
+  return { begin, end, beginCommit, collectFiber, endCommit, takeTrace };
+};
+
+// ---------------------------------------------------------------------------
+// Scanner: a single global commit hook driving a canvas of re-render outlines.
+// ---------------------------------------------------------------------------
 
 interface ScanOutline {
   nameCounts: Map<string, number>;
-  // Cached label, recomputed only on commit (not per frame) to keep the
-  // render loop allocation-free.
   label: string;
   lastRenderTimestamp: number;
 }
 
-export interface ScannerController {
+interface ScannerController {
   start: () => void;
   stop: () => void;
   isScanning: () => boolean;
@@ -47,9 +211,6 @@ export interface ScannerController {
 
 let handleCommit: (root: FiberRoot) => void = () => {};
 
-// Tracks whether bippy is instrumenting a React renderer. Scanning is
-// meaningless without one, so the toolbar hides the scan button until this
-// flips true (on instrumentation activating or the first React commit).
 let hasActiveInstrumentation = isInstrumentationActive();
 const instrumentationListeners = new Set<() => void>();
 
@@ -59,18 +220,16 @@ const markInstrumentationActive = (): void => {
   for (const listener of instrumentationListeners) listener();
 };
 
-export const isScanAvailable = (): boolean => hasActiveInstrumentation;
+const isScanAvailable = (): boolean => hasActiveInstrumentation;
 
-export const onScanAvailable = (listener: () => void): (() => void) => {
+const onScanAvailable = (listener: () => void): (() => void) => {
   instrumentationListeners.add(listener);
   return () => instrumentationListeners.delete(listener);
 };
 
 if (typeof window !== "undefined") {
-  // Raw instrument() (not secure()): secure tears the commit hook down if a
-  // React renderer isn't detected within its ~100ms install-check window, which
-  // breaks scanning on apps that mount React after react-grab's script loads.
-  // React core already wraps onCommitFiberRoot in try/catch, so this is safe.
+  // Not secure(): it tears the hook down if React isn't detected within ~100ms,
+  // which breaks apps that mount React after react-grab's script loads.
   instrument({
     name: "react-grab",
     onActive: markInstrumentationActive,
@@ -81,7 +240,7 @@ if (typeof window !== "undefined") {
   });
 }
 
-export const createScanner = (): ScannerController => {
+const createScanner = (): ScannerController => {
   let isScanning = false;
   let canvas: HTMLCanvasElement | null = null;
   let context: CanvasRenderingContext2D | null = null;
@@ -93,8 +252,6 @@ export const createScanner = (): ScannerController => {
   const outlines = new Map<HTMLElement, ScanOutline>();
   const recorder = createScanTraceRecorder();
 
-  // Set once per commit so the per-fiber recorder reads a single clock value
-  // instead of calling performance.now() inside the hot fiber walk.
   let currentCommitTimestamp = 0;
 
   const didNearestCompositeAncestorRender = (fiber: Fiber): boolean => {
@@ -242,8 +399,6 @@ export const createScanner = (): ScannerController => {
     scheduleFrame();
   };
 
-  // The recorder's trace is intentionally preserved after stop() so the caller
-  // can read it via takeTrace(); the next start() clears it via recorder.begin().
   const stop = (): void => {
     if (!isScanning) return;
     isScanning = false;
@@ -265,4 +420,60 @@ export const createScanner = (): ScannerController => {
     isScanning: () => isScanning,
     takeTrace: recorder.takeTrace,
   };
+};
+
+// ---------------------------------------------------------------------------
+// Controller: reactive scan state, clipboard copy on stop, "copied" toast.
+// ---------------------------------------------------------------------------
+
+export interface ScanController {
+  isScanAvailable: Accessor<boolean>;
+  isScanning: Accessor<boolean>;
+  // Fresh token per copy so the toolbar can key (and replay) its toast.
+  scanCopiedToken: Accessor<number | null>;
+  toggle: () => void;
+  stop: () => void;
+}
+
+export const createScanController = (): ScanController => {
+  const scanner = createScanner();
+  const [scanAvailable, setScanAvailable] = createSignal(isScanAvailable());
+  onCleanup(onScanAvailable(() => setScanAvailable(true)));
+  const [isScanning, setIsScanning] = createSignal(false);
+  const [scanCopiedToken, setScanCopiedToken] = createSignal<number | null>(null);
+  let scanCopiedCount = 0;
+  let scanCopiedTimeout: ReturnType<typeof setTimeout> | undefined;
+
+  const flashCopied = () => {
+    scanCopiedCount += 1;
+    setScanCopiedToken(scanCopiedCount);
+    clearTimeout(scanCopiedTimeout);
+    scanCopiedTimeout = setTimeout(() => setScanCopiedToken(null), FEEDBACK_DURATION_MS);
+  };
+
+  const stop = () => {
+    if (!scanner.isScanning()) return;
+    scanner.stop();
+    setIsScanning(false);
+  };
+
+  const toggle = () => {
+    if (scanner.isScanning()) {
+      stop();
+      const trace = scanner.takeTrace();
+      if (trace && copyContent(serializeScanTrace(trace), { componentName: "ReactGrabScan" })) {
+        flashCopied();
+      }
+      return;
+    }
+    scanner.start();
+    setIsScanning(true);
+  };
+
+  onCleanup(() => {
+    clearTimeout(scanCopiedTimeout);
+    stop();
+  });
+
+  return { isScanAvailable: scanAvailable, isScanning, scanCopiedToken, toggle, stop };
 };
