+++ packages/react-grab/src/styles.css
@@ -337,8 +337,6 @@
   will-change: transform, opacity;
 }
 
-/* Hold, then fade+blur out - matches the selection label's dismissal. The
-   hold duration comes from the animation-delay set inline on the element. */
 @keyframes rg-copied-pill-out {
   from {
     opacity: 1;
